hdl/devices/lime_dac.hdl/lime_dac.xml: add ZLM on input (maybe not supported in code?)
hdl/devices/sdp_receive.hdl/sdp_receive.vhd: split remote/local buffer counts - asymmetry
hdl/devices/sdp_receive.hdl/sdp_receive_dma.vhd: ditto
hdl/platforms/zed/Makefile: re-enable configurations
runtime/application/include/OcpiApplication.h: support for xferrole and buffercount in ocpirun
runtime/application/src/OcpiApplication.cxx: ditto
runtime/application/src/ocpirun_main.cxx: ditto
runtime/dataplane/rdma_drivers/dma/src/DtDmaXfer.cxx: better message
runtime/hdl-support/src/ocpinft_main.c: merge with develop
runtime/library/src/OcpiLibraryAssembly.cxx: feed params down into assembly parsing
runtime/util/assembly/include/OcpiUtilAssembly.h: ditto
runtime/util/assembly/src/OcpiUtilAssembly.cxx: ditto
tools/cdk/include/hdl/hdl-platform.mk: skip biulding platforms not in HdlPlatforms
tools/cdk/include/rcc/rcc-worker.mk: use OCPI_USE_TOOL_MODES, improve indents fix -A name

+++ os/interfaces/include/OcpiOsEther.h
@@ -38,7 +38,7 @@
 #ifndef OCPIOSETHER_H__
 #define OCPIOSETHER_H__
 
-#include <stdint.h>
+#include <inttypes.h>
 #include <memory.h>
 #include <stdio.h>
 #include <string>
@@ -127,7 +127,7 @@ namespace OCPI {
 	uint8_t source[Address::s_size];
 	uint8_t payload[MaxPacketSize];
       };
-      // A socket that sends and receives L2 layer poackets.
+      // A socket that sends and receives L2 layer packets.
       // Note that receives are done into internal buffers and it is assumed that
       // the packet buffer can be reused on the next receive call.
       // This socket is for communicating to a simple adress,
@@ -179,7 +179,7 @@ namespace OCPI {
       class IfScanner {
 	uint64_t m_opaque[4];
 	bool     m_init;
-	unsigned m_index;
+	size_t   m_index;
       public:
 	// Initialize the scanner
 	// err is set if there is an error
